0010 Testing for new model that image can be copied and deleted

+++ Mosaic/Mosaic.WebUI/Mosaic.WebUI/Models/ViewImageModel.cs
@@ -8,28 +8,41 @@ namespace Mosaic.WebUI.Models
 {
     public class ViewImageModel
     {
-        private readonly string ROOT_IMAGE_FOLDER = "~\\images\\temp\\";
+        public string CopyPath { get; set; }
         public string FilePath { get; set; }
         public string Error { get; set; }
 
+        public ViewImageModel(string copyPath = "~\\images\\temp\\")
+        {
+            CopyPath = copyPath;
+        }
+
         public void CopyImage(string fileToCopy)
         {
-            if (File.Exists(fileToCopy))
+            if (Directory.Exists(CopyPath))
             {
-                var newfile = ROOT_IMAGE_FOLDER + Path.GetFileName(fileToCopy);
-                File.Copy(fileToCopy, newfile);
-                FilePath = newfile;
+                if (File.Exists(fileToCopy))
+                {
+                    var newfile = CopyPath + Path.GetFileName(fileToCopy);
+                    File.Copy(fileToCopy, newfile);
+                    FilePath = newfile;
+                }
+                else
+                {
+                    Error = "File to copy does not exist";
+                }
             }
             else
             {
-                Error = "File to copy does not exist";
+                Error = "Image directory does not exist";
             }
+            
         }
 
         // Deletes the current image that has been copied previous
         public void DeleteImage()
         {
-            var di = new DirectoryInfo(ROOT_IMAGE_FOLDER);
+            var di = new DirectoryInfo(CopyPath);
             foreach (FileInfo file in di.GetFiles())
             {
                 file.Delete();
